Javadoc cleanup and simplification of the simple system
