+++ library/block/block_content_registry/src/main/java/org/quiltmc/qsl/block/content/registry/impl/BlockContentRegistriesInitializer.java
@@ -36,7 +36,7 @@ import net.minecraft.item.ShovelItem;
 import org.quiltmc.loader.api.ModContainer;
 import org.quiltmc.qsl.base.api.entrypoint.ModInitializer;
 import org.quiltmc.qsl.block.content.registry.api.BlockContentRegistries;
-import org.quiltmc.qsl.block.content.registry.api.EnchantingBoosters;
+import org.quiltmc.qsl.block.content.registry.api.enchanting.ConstantBooster;
 import org.quiltmc.qsl.block.content.registry.api.FlammableBlockEntry;
 import org.quiltmc.qsl.block.content.registry.api.ReversibleBlockEntry;
 import org.quiltmc.qsl.registry.attachment.api.RegistryEntryAttachment;
@@ -80,7 +80,7 @@ public class BlockContentRegistriesInitializer implements ModInitializer {
 		)), BlockContentRegistries.WAXABLE);
 		addMapToAttachment(initialFlammableBlocks, BlockContentRegistries.FLAMMABLE);
 
-		BlockContentRegistries.ENCHANTING_BOOSTERS.put(Blocks.BOOKSHELF, new EnchantingBoosters.ConstantBooster(1f));
+		BlockContentRegistries.ENCHANTING_BOOSTERS.put(Blocks.BOOKSHELF, new ConstantBooster(1f));
 
 		resetMaps();
 		ResourceLoaderEvents.END_DATA_PACK_RELOAD.register((server, resourceManager, error) -> resetMaps());
