+++ library/block/block_content_registry/src/testmod/java/org/quiltmc/qsl/block/test/BlockContentRegistryTest.java
@@ -19,7 +19,10 @@ package org.quiltmc.qsl.block.test;
 import java.util.Optional;
 
 import com.mojang.serialization.Codec;
-import org.quiltmc.qsl.block.content.registry.api.EnchantingBoosters;
+import org.quiltmc.qsl.block.content.registry.api.enchanting.ConstantBooster;
+import org.quiltmc.qsl.block.content.registry.api.enchanting.EnchantingBooster;
+import org.quiltmc.qsl.block.content.registry.api.enchanting.EnchantingBoosterType;
+import org.quiltmc.qsl.block.content.registry.api.enchanting.EnchantingBoosters;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -57,10 +60,10 @@ public class BlockContentRegistryTest implements ModInitializer {
 				new OxidizableBlock(Oxidizable.OxidizationLevel.UNAFFECTED, AbstractBlock.Settings.copy(Blocks.IRON_BLOCK)),
 				BlockContentRegistries.OXIDIZABLE, new ReversibleBlockEntry(Blocks.IRON_BLOCK, false));
 
-		BlockContentRegistries.ENCHANTING_BOOSTERS.put(Blocks.IRON_BLOCK, new EnchantingBoosters.ConstantBooster(3f));
-		BlockContentRegistries.ENCHANTING_BOOSTERS.put(Blocks.DIAMOND_BLOCK, new EnchantingBoosters.ConstantBooster(15f));
-		BlockContentRegistries.ENCHANTING_BOOSTERS.put(Blocks.NETHERITE_BLOCK, new EnchantingBoosters.ConstantBooster(100f));
-		BlockContentRegistries.ENCHANTING_BOOSTERS.put(Blocks.OAK_PLANKS, new EnchantingBoosters.ConstantBooster(0.25f));
+		BlockContentRegistries.ENCHANTING_BOOSTERS.put(Blocks.IRON_BLOCK, new ConstantBooster(3f));
+		BlockContentRegistries.ENCHANTING_BOOSTERS.put(Blocks.DIAMOND_BLOCK, new ConstantBooster(15f));
+		BlockContentRegistries.ENCHANTING_BOOSTERS.put(Blocks.NETHERITE_BLOCK, new ConstantBooster(100f));
+		BlockContentRegistries.ENCHANTING_BOOSTERS.put(Blocks.OAK_PLANKS, new ConstantBooster(0.25f));
 		BlockContentRegistries.ENCHANTING_BOOSTERS.put(Blocks.REDSTONE_WIRE, new EnchantingBlockStateBooster());
 
 		ServerWorldTickEvents.START.register((server, world) -> {
@@ -80,8 +83,9 @@ public class BlockContentRegistryTest implements ModInitializer {
 		});
 	}
 
-	private record EnchantingBlockStateBooster() implements EnchantingBoosters.EnchantingBooster {
-		public static EnchantingBoosters.EnchantingBoosterType TYPE = EnchantingBoosters.register(new Identifier(MOD_ID, "block_state_booster"), Codec.unit(EnchantingBlockStateBooster::new));
+	private record EnchantingBlockStateBooster() implements EnchantingBooster {
+		public static EnchantingBoosterType TYPE = EnchantingBoosters.register(new Identifier(MOD_ID, "block_state_booster"),
+				new EnchantingBoosterType(Codec.unit(EnchantingBlockStateBooster::new), Optional.of(new EnchantingBlockStateBooster())));
 		@Override
 		public float getEnchantingBoost(World world, BlockState state, BlockPos pos) {
 			if (!state.contains(Properties.POWER)) {
@@ -92,7 +96,7 @@ public class BlockContentRegistryTest implements ModInitializer {
 		}
 
 		@Override
-		public EnchantingBoosters.EnchantingBoosterType getType() {
+		public EnchantingBoosterType getType() {
 			return TYPE;
 		}
 	}
